Move Sha1HashCode into its own package.
Summary:
Previously, it was in `com.facebook.buck.rules`, which did not make
a lot of sense, as it was already in its own `java_library()` rule.
Now it is in its own package, `com.facebook.buck.util.sha1`, which
will make it easier to depend on.

The only benefit of the old location was that its fields
(`firstFourBytes`, `nextEightBytes`, `lastEightBytes`)
could be package-private so that `RuleKeyBuilder` could access them.
Now those fields are public, though it is of little concern,
as they are immutable.

Other than the change to the visibility of the fields, all of the other changes were
performed mechanically by IntelliJ (or `buck autodeps`).

Test Plan: CI

Reviewed By: Coneko

fbshipit-source-id: 52128f6