Do not throw when an import has no capital letters because there are third-party situations that we cannot control.
Summary:
I tried running `buck autodeps` on a larger Android codebases with
a wider array of third-party dependencies and I ran into this issue.
It makes me wish that types-must-start-with-uppercase were mandated
at the language level...

Test Plan:
There was an existing test case for this in `JavaFileParserTest`.
It has been updated to reflect the change in policy.

Reviewed By: Coneko

fbshipit-source-id: fb311cb