Refactor logic for extracting dependency info for Java rules into its own method.
Summary:
This extracts out a portion of `JavaDepsFinder.findDepsForBuildFiles()` into its
own method, `DepsForBuildFiles.findDependencyInfoForGraph()`. To make this
work, the relevant data returned by `findDependencyInfoForGraph()` must be
encapsulated in its own type, `DepsForBuildFiles.DependencyInfo`.

The bulk of the original version of `findDepsForBuildFiles()` is now in encompassed
by a private method of same name. It takes a `DependencyInfo` as its primary argument
and no longer needs the `TargetGraph` as a parameter because
`findDependencyInfoForGraph()` does all of the graph processing it needs.

The new `findDependencyInfoForGraph()` method is used in an upcoming revision to
support `buck autodeps --suggest-partition`.

Test Plan: This is a refactoring, so it is covered by the existing tests.

Reviewed By: Coneko

fbshipit-source-id: 00a146b