Fix runtime dependencies for apk_genrule targets
Summary:
An apk_genrule target isn't providing the manifest directly,
but refers to the manifest of its dependency instead. That means that
dependency is a runtime dependency of the android instrumentation test
rule as well.

Test Plan:
Ran instrumentation test, made sure I was uploading to the
cache. Removed all files in `buck-out/`. Ran the test again. Confirmed
all binaries came from cache and that the test now succeeded instead of
failed with an error in buck.

Reviewed By: yiding

fbshipit-source-id: a4dea10