Throw correct exception when BuildTargetTypeCoercer encounters non-string
Summary:
When encountering a non-string, BuildTargetTypeCoercer threw an
IllegalArgumentException. Since this is actually a user error, we should throw
CoerceFailedException instead.

This allows caller to properly emit an error message to the user.

Test Plan:
pass a list to a field expecting a build target, see that I get an informative
error message instead of an IllegalArgumentException

Reviewed By: andrewjcg

fbshipit-source-id: c4fa23c