diff_rulekeys: less noisy output
Summary:
If the lists do not contain the same set of elements we will just
output the elements which are in one set but not the other instead of comparing
lists on an index basis. Before this change the mismatch between the lists
`[a,b,c]` and `[b,c,d]` would be shown as `[-a,+b],[-b,+c],[-c,+d]`, with the
change we show it as `[-a, +d]`.

Test Plan: new unittest

Reviewed By: michath

fbshipit-source-id: f1c2bb6