python: exclude transitive preload deps from omnibus merging
Summary:
Normally, preloaded deps lists are carefully constructed to minimize
the amount of code that gets preloaded.  With omnibus linking, since
we merge all deps into a giant DSO, this can sometimes lead to preload
deps pulling in most of the world if one of it's transitive deps are
included in the omnibus library.

This diff explicitly excludes the transitive deps of preloaded deps
from omnibus linking to avoid this.

Test Plan: added unittest

Reviewed By: beefon

fbshipit-source-id: 61e9ee2