react-native: used manifest-based caching for packager
Summary:
This ports the existing react native packager code to use manifest-based
caching rather than the older ABI rule interface (which only works locally).
Currently this calls into the packager twice: once to actual perform the
task and another time to grab the dep files.  Before we use this we should
add a mode to the packager which enables it to do both tasks in a single
run.

Test Plan: ran js tests

Reviewed By: Coneko

fbshipit-source-id: 85c3703