Correctly handle lists of variable length in diff_rulekeys
Summary:
The script would truncate the longer list of values when comparing leading
to spurious change detection (`[1,2,3]` vs `[2,3]` would be compared as
`[1,2]` vs `[2,3]` resulting in a confusing `-[1], +[3]` diff).
This was caused by the incorrect use of the `zip` function in Python,
which was replaced by `map(None, ...)`.

Test Plan: added new unittests

Reviewed By: michath

fbshipit-source-id: a75a251