buck query: batch preloading target patterns in multi-query
Summary:
Previously, a multi-query command would only preload the target patterns
from the queries one at a time, which meant we never took advantage of
the full parallelism of parsing.  This diff changes the interface and
semantics of the parser's target spec resolution methods, so that it
doesn't merged resolved targets into a single set, but rather as a list
of sets that correspond to the order of the target specs passed into the
method.

The multi-query command has also been changed so that it first collects
up all target patterns from all queries and then preloads them all in
one go, which can be significantly more efficient.

Test Plan:
- CI
- Added test for new parser interface

Reviewed By: Coneko

fbshipit-source-id: 4335a3a