Force AbstractNetworkCache to write its temp files under buck-out/ rather than tmp/.
Summary:
This was causing issues for [Eden](https://github.com/facebookexperimental/eden) because
writes are fairly expensive as compared to reads. Currently, we mitigate this issue by bind-mounting
the `buck-out` directory in EdenFS, so we are less concerned about the I/O that happens there.

It turns out that `AbstractNetworkCache` was using the `tmp` directory in the project root as its own
scratch directory, so this fixes it to use `buck-out/tmp` instead. Eden notwithstanding, letting Buck do
writes outside of `buck-out` is not kosher. I suspect this went unnoticed because in practice:

1. This only happens if you are using `http` or `thrift_over_http` as a cache mode.
2. `AbstractNetworkCache` attempts to clean up the file in `tmp` after the artifact is uploaded.

Test Plan: CI

Reviewed By: Coneko

fbshipit-source-id: b3d4aab