Make AndroidNativeLibsPackageableGraphEnhancer.generateStripRules static
Summary:
I'm updating this method to make it share rules between apps.  Making it
static is an optional step to reduce the risk of details of the app
leaking into the generated rules (by making all data passed in
explicitly).  LMK if you think this is a bad idea or waste of space.

Test Plan: CI.

Reviewed By: andrewjcg, asp2insp

fbshipit-source-id: 3d4dd97