Fix deadlock on shutdown.
Summary:
In addition to buck, both  nuprocess and nailgun install shutdown hooks.
These are threads which run when the JVM is about to exit but are still
able to invoke the unhandled exception handler when an exception is thrown
from within any of those shutdown hook threads. The buck handler calls
System.exit at the end. Unfortunately the System.exit implementation does
all of the operations under a lock (both waiting for the shutdown hooks
and checking if we're already shutting down to early-out) meaning that
we're able to easily get into a deadlock.
This change breaks the cycle by moving the System.exit call to happen
on a dedicated thread.

Test Plan: added new test

Reviewed By: Coneko

fbshipit-source-id: 371f4eb