Set read and write timeouts for http artifact store clients
Summary:
The http fetcher has a read timeout, but the storer only has a connect timeout.

This sometimes causes builds to hang on a store, as for whatever reason the
remote side never responds and the connection never times out.

This might affect storing large objects, but I'm not sure that's a problem in practice...

Test Plan: visualscrutiny

Reviewed By: ruibm, Coneko

fbshipit-source-id: 45f968e