Introducing ProjectFilesystemDelegate
Summary:
This introduces a new parameter to `ProjectFilesystem`: `ProjectFilesystemDelegate`.
The purpose of the delegate is explained in the interface overview:

> Delegate that a {link ProjectFilesystem} can use to leverage a specialized implementation of
> certain filesystem operations, tailored to the underlying filesystem. Use of the delegate is
> often motivated by performance reasons.

This revision does not introduce any actual use of delegation yet, but it sets
things up so we can create an `EdenProjectFilesystemDelegate` in a follow-up revision.

In general, the right way to create a `ProjectFilesystemDelegate` is via
`ProjectFilesystemDelegateFactory#newInstance(Path)`. Indeed, this revision
makes this the default behavior when creating a `ProjectFilesystem`.
(There are too many call sites to `new ProjectFilesystem` to make it worth it
to require `ProjectFilesystemDelegate` to be specified explicitly in all cases,
particularly when there is a sensible default in the overwhelming majority of cases.)

That said, it is possible to pass a specific `ProjectFilesystemDelegate` from above.
`FakeProjectFilesystem` takes advantage of this to ensure that a
`DefaultProjectFilesystemDelegate` is always used for testing.

Test Plan: CI

Reviewed By: marcinkosiba

fbshipit-source-id: cd4659a