Changes to ActionGraphCache check mechanism
Summary:
Made the call synchronous to the main thread.
Possible mismatch causes a Runtime exception.
Updated the documentation.
Made the value be true/false.

Note: Immutable SampleRate is not currently needed but I decided not to delete it since it could
be useful in a future. If we decide to remove it I can do it at another diff.

Reviewed By: marcinkosiba

fbshipit-source-id: 633949f