Use APPEND_TO_ZIP instead of OVERWRITE_EXISTING for buck rage reports
Summary:
OVERWRITE_EXISTING is inefficient and we should avoid that mode. This diff makes it one step closer to killing it completely.
Furthermore, when we bundle files to a buck rage report, we should not overwrite the entries. That would give us incorrect perception of the actual directory state. Instead we should use APPEND_TO_ZIP to get all entries, or THROW_EXCEPTION if we are sure there are no duplicated entries.

Unfortunately, due to how `DefaultDefectReporter.addFilesToArchive` works, we can get duplicate entries even if all the paths are unique because we are stripping off the `.` of hidden files. So the files `.blah` and `blah` would both end up with the same entry name. Another possible collision is if `report.json` is in the list of files to zip as we create an entry with such name from an in-memory structure and not from an actual file. If such unlikely pathological scenario happens, it is better to have both files in the report than only the last one.

Test Plan:
buck rage
https://our.intern.facebook.com/intern/buck/rage/524538301084026

Reviewed By: marcinkosiba

fbshipit-source-id: c8be1c6