Expose parsing status to console
Summary:
The goal is to provide visual feedback to the user about how optimal the parsing stage went.
Previous work has introduced PerformanceEvents that the console subscribes and this way
we can track some basic events that could explain a slow parsing stage or an optimal result.

The events tracked are:
- Watchman events: Overflow, file addition, file removal.
- Action graph cache miss or hit
- Invalidation of parsing caches due to a symlink introduction.

ps. The emojis are just for fun, if the console can not reproduce them they are replaced with
spaces.

Test Plan:
Examples of output:

[-] PROCESSING BUCK FILES...FINISHED 0.3s 🐇
[-] PROCESSING BUCK FILES...FINISHED 1.0s [100%] 🐌  (Files were added or removed)
[-] PROCESSING BUCK FILES...FINISHED 1.4s [100%]  🐌  (Watchman overflow event)

Reviewed By: marcinkosiba

fbshipit-source-id: 7f25541