Make it possible to run a build from a serialized state dump.
Summary:
This hooks up all of the distributed build serialization to the
`buck build --distributed --distributed-state-file` command and
makes it possible to run a build solely from a state snapshot.
Currently the snapshot is entirely self-contained (and therefore
fairly big as it includes all of the sources and binary data) and
not all of the rules work with this setup (anything that touches
the disk during ActionGraph creation, like resource validation
for java_library or android_resources are broken).

Test Plan:
new tests, manually:
  buck build --distributed --distributed-state-dump=/tmp/dump.bin src/com/facebook/buck/shell/...
  mkdir /tmp/test; cd /tmp/test; touch .buckconfig
  buck build --distributed --distributed-state-dump=/tmp/dump.bin src/com/facebook/buck/shell/...

Reviewed By: ruibm

fbshipit-source-id: 0fd7941