Collapse multiple / into one
Summary:
A number of people incorrectly use `glob()` in BUCK files like this:

  some_library(
    srcs=glob(['path/to//src/**/*.java']),
  )

The double-slash typo happens to be treated the same as a single slash
with Python `glob()` (and `pathlib`), but it does not currently work with
Watchman glob (which uses the `wildmatch.c` library, which always fails
to match any pattern with multiple slashes).

I sent a fix upstream to Watchman:

https://github.com/facebook/watchman/pull/310

which sid0 kindly merged, but it will take time to deploy a fix, so
this also works around the bug on the Buck side.

Test Plan:
`buck build //path/to:target` where `srcs` included
a glob with multiple slashes

Reviewed By: ryu2

fbshipit-source-id: 2828a38