Defensively copy memoized cache results
Summary:
A while back, I added a memoizing cache to the Watchman glob
function.

Unfortunately, a number of clients mutate the results of `glob()`,
which was poisoning the cache.

This adds a defensive copy to the result returned from the memoizing
cache, so we can rest easy that people won't mutate the result.

(I tried changing it to a tuple, but people commonly append
lists with the result of `glob()`, and you cannot append lists and
tuples in Python.)

Test Plan:
`buck build //path/to:target` in repo which mutated
result of `glob()`

Reviewed By: yiding

fbshipit-source-id: 1fd0aa1