Allow halide_library() to override the generated function name
Summary:
I'm working with an app that relies on two halide libraries
that are essentially the same (except they're intended to be included
on different platforms).  The two targets have different names, but
need to generate functions with identical names to make it easier on the
test app I'm working with.

Test Plan: Get farther building a Halide-dependent test app on Linux.

Reviewed By: Coneko

fbshipit-source-id: 1600a3c