Remove FileListBuilder after use
Summary:
The new dep files for java code utilizes a `DependencyFileInputListBuilder`
to track class usages when compiling a java library.
This list builder keeps the dependency files cached in memory, but the builder (along with that cache)
is retained throughout the entire build process.
This memory usage scales with the total number of java files touched by the entire build, so it's
worth eliminating.

This change now allows the cache to be read from the persisted disk location, bringing it inline
with the C++ implementation.

Test Plan: All unit tests pass.

Reviewed By: yiding

fbshipit-source-id: 64a9bc6