Account for non-String ssid in MacWifiSsidFinder
When attempting to build on the our corporate wifi network, I get the following error:

```
[2016-07-27 06:40:00.651][error][command:3b0248e1-8685-4e49-bf44-591897aa465a][tid:18][com.facebook.buck.util.environment.MacWifiSsidFinder] Fetched SSID, but got unexpected non-string type (got: Google).
```

`Google` in this case is the SSID name, not the type. After updating the code to return the class name and running again, I get the following:

```
[2016-07-27 06:42:08.760][error][command:ad511766-4861-4615-98ee-9f8130ee5387][tid:18][com.facebook.buck.util.environment.MacWifiSsidFinder] Fetched SSID, but got unexpected non-string type (got: ca.weblite.objc.Proxy).
```

So it appears that in some (I’m assuming rare) cases, the call to `ssid` is resulting in a `Proxy` object being returned instead of a `String`.

Calling `toString()` solves this issue, but I’m also curious to know why this would happen in the first place.