Account for non-String ssid in MacWifiSsidFinder
Summary:
When attempting to build on the our corporate wifi network, I get the following error:

```
[com.facebook.buck.util.environment.MacWifiSsidFinder] Fetched SSID, but got unexpected non-string type (got: Google).
```

`Google` in this case is the SSID name, not the type. After updating the code to return the class name and running again, I get the following:

```
[com.facebook.buck.util.environment.MacWifiSsidFinder] Fetched SSID, but got unexpected non-string type (got: ca.weblite.objc.Proxy).
```

So it appears that in some (I’m assuming rare) cases, the call to `ssid` is resulting in a `Proxy` object being returned instead of a `String`.

Calling `toString()` solves this issue, but I’m also curious to know why this would happen in the first place.

This is running on 10.11.6 (15G31)
Closes https://github.com/facebook/buck/pull/819

Reviewed By: ryu2

fbshipit-source-id: 4555e3a