Removed extraneous output classpath entry building
Summary:
Output Classpath Entries consumes a significant
amount of CPU and memory to build and memoize a mapping of `JavaLibrary->{Path}`
representing the transitive classpath of a given `JavaLibrary`. This information
is never used. Every call site wraps the call with a variation on
`ImmutableSet.copyOf(library.getOutputClassPathEntries().values())`
This means that we store far too many entries on each node (since we copy and double
the transitive entries from the level below) and we spend a lot of time and effort
iterating and deduplicating.

This change removes the mapping of entries and moves to a simple `ImmutableSet<Path>`.

The last substantial modification to this code was in 2013, so I don't believe this
 info/functionality will be necessary any time soon.

Test Plan: all associated unit tests pass.

Reviewed By: yiding

fbshipit-source-id: ec873f8