Add a scaffold for merging native libraries in Android builds
Summary:
We need a per-app way to merge native libraries to reduce our DSO count.
This sets up the basic scaffold with the config threaded through and an
integration point, similar to how we do relinking.  I'm sending this as
a separate diff to reduce the noise in the main diffs.

Test Plan: CI

fbshipit-source-id: 40ecf1d