Update DefaultFileHashCache to use ProjectFilesystem#computeSha1()
Summary:
This appears to be the main place where the SHA-1 of a file was being computed
that did not go through `ProjectFilesystem`. This should consolidate things.

Test Plan:
1. CI
2. Put a `println()` in `ProjectFilesystem.computeSha1()` when doing `NO_BUCKD=1 buck build -v 0 buck | grep -E '\.java$'` and verify it requests the expected hashes (after doing an `rm -rf buck-out`).

I took the output from my `println()` experiment, filtered to include only paths that started with `src/` and then sorted and wrote the output to `/tmp/output.txt`. I then got the list of files in the `src` folder via:

    find src -name \*.java | grep -v intellij/plugin > /tmp/all.txt

Here is the output of `diff /tmp/output.txt /tmp/all.txt`:

```
16a17
> src/com/facebook/buck/android/agent/AgentMain.java
117a119,124
> src/com/facebook/buck/android/support/exopackage/ApplicationLike.java
> src/com/facebook/buck/android/support/exopackage/DefaultApplicationLike.java
> src/com/facebook/buck/android/support/exopackage/ExopackageApplication.java
> src/com/facebook/buck/android/support/exopackage/ExopackageDexLoader.java
> src/com/facebook/buck/android/support/exopackage/ExopackageSoLoader.java
> src/com/facebook/buck/android/support/exopackage/SystemClassLoaderAdder.java
162a170
> src/com/facebook/buck/apple/clang/PrintHeaderMap.java
234a243
> src/com/facebook/buck/apple/xcode/xcspec/SDKSettings.java
259a269
> src/com/facebook/buck/artifact_cache/HybridPayloadGenerator.java
793a804
> src/com/facebook/buck/jvm/java/abi/ApiStubber.java
1491a1503,1506
> src/com/facebook/buck/tools/dxanalysis/DxAnalysisMain.java
> src/com/facebook/buck/tools/dxanalysis/MutabilityAnalyzer.java
> src/com/facebook/buck/tools/dxanalysis/RegisterSpecAnalyzer.java
> src/com/facebook/buck/tools/dxanalysis/StaticStateAnalyzer.java
1514a1530
> src/com/facebook/buck/util/collect/ArrayIterable.java
```

In other words, there are only 16 `.java` files in the `src/` folder whose SHA-1 is not requested when doing `buck build buck`
(not counting the IntelliJ plugin, which we don't expect to build as part of `buck build buck`). Auditing these 16 files, it is
expected that they are not sourced by `buck build buck`. (More importantly, it appears that `SDKSettings.java` and
`ArrayIterable.java` can be deleted, as they are only referenced from their unit tests and nowhere else. Some files,
such as `PrintHeaderMap.java`, `HybridPayloadGenerator.java`, and `ApiStubber.java` are used by standalone `java_binary()` rules.)

Reviewed By: yiding

fbshipit-source-id: 3550668