Make sure readonly paths are scoped to current project
Summary:
I'm experimenting with cross-project dependencies. One of the projects
with a readonly section was failing the build because the paths were referenced
to the host project being built, not the dependent project.

Solution: make sure that parsing read-only paths goes through the same
code as any other path.  This makes sense for read-only paths, since any
path in a project ought to be scoped to that project's filesystem.

Test Plan: Successfully build app.

Reviewed By: Coneko

fbshipit-source-id: ab7cec6