Change the default temp directory for java_test to not be in buck-out.
Summary:
This was originally changed to be in buck-out in #16, to avoid having the
tests pollute the system temporary directory. However none of the other
tests do this. Also features that were added in the meantime allow this
to be specified in java_test's vm_args and env attributes.

Removing this also allows us to remove the temp_dir_env_vars setting, which
was only added to allow disabling this feature.

Test Plan: CI

Reviewed By: bhamiltoncx

fbshipit-source-id: f4b038e