Allow overriding multiple host platforms
Summary:
When building a cross-platform project for Linux and Mac
that happens to override both [cxx#linux-x86_64] and [cxx#macosx-x86_64],
the MacOSX build complained about being unable to find the [cxx#linux-x86_64] platform.

This is reasonable; the Linux platform isn't meant to be used when building
on MacOSX. However, this isn't an error condition, so the build should continue.

Test Plan: Unit test.

Reviewed By: Coneko

fbshipit-source-id: 87987da