buck query: fix `rdeps` when used with symlinks/autodeps
Summary:
The buck query implementation calls into the parser several different
times to parse target nodes and graphs.  Each call into the parser
creates and destroys a new per-built state object which invalidates
nodes in the parser cache when it gets cleaned up (e.g. if they were
found to be past symlinks).  This means the query command could
potentially re-parse the same target node across these separate calls
into the parser.

Normally, this would only affect performance.  However, since target
nodes are now compared throughout the codebase using instance equality,
we can end up with multiple nodes for the same build target which are
not "equal".  This can result in internal crashes, as some parts of
the code get confused by this scenario.

Ideally, we'd use maps to represent the target graph, so we could use
the build target as a key and not rely on the equality implementation.
However, in this case, since this was also affecting performance, this
diff adds a new interface to the parser object, so that clients can
get a handle that maintains persistent state across multiple calls.
This should be more efficient and resolve the crashes we've been seeing.

Test Plan: added integration test which repros the crash

Reviewed By: marcinkosiba

fbshipit-source-id: d0c151c