Check for some errors in NativeLibraryMergeEnhancer
Summary:
Flag attempting to merge a library-as-asset with a library that can't be
used as an asset (this seems like it would violate expectations) and
attempting to merge one library into two different shared objects.

This requires building up some of the structures that we'll need to
prepare actual library merging.

Test Plan: CI

fbshipit-source-id: c23f43b