Make RelinkerRule's RuleKey stable.
Summary:
When running the scripts/diff_rulekeys.py script I found that the
ordering of the list varies across different machines. Given the use of the
list (read it and merge all symbols into a set) reordering and duplicate
removal should have no impact on the result.

Test Plan: CI

Reviewed By: Coneko

fbshipit-source-id: 8e6d9d6