Use EdenProjectFilesystemDelegate when root is subdirectory of Eden mount.
Summary:
We may have cases where an Eden mount contains a `.buckconfig` file that is
not in its root. When this happens, when `EdenProjectFilesystemDelegate` makes
calls to Eden's Thrift API, it must prefix relative Buck project paths so that
they are relative to the root of the Eden mount.

For good measure, I also updated the Eden Thrift JARs.

Test Plan:
* Updated unit tests.
* Manually tested using Buck in an Eden mount with a `.buckconfig` in a subdirctory.
* `ant java-test -Dtest.class=EdenMountTest` and `ant java-test -Dtest.class=EdenClientTest`

Reviewed By: marcinkosiba

fbshipit-source-id: 23e2190