Remove third-party/java/eden/thrift-java.jar that was inadvertently added.
Summary:
This was just added as part of D3709597, but it should have been added as
`eden.jar` to overwrite the existing Thrift bindings for Eden. There are no
references to `thrift-java.jar`, so adding this file had no effect.

Presently, I'm planning to put up a new revision that is going to include a new
version of `eden.jar` with an API change, so we might as well just delete this
ASAP.

Test Plan: CI

Reviewed By: Coneko

fbshipit-source-id: af4fe78