Make EdenProjectFilesystemDelegate.computeSha1() retry symlinks appropriately.
Summary:
The Thrift API for Eden fails if the path passed to `getSHA1()` is a symlink.
This adds logic to Buck to try to recover from this in a graceful way.

This upgrades us from jimfs-1.0 to jimfs-1.1 to pick up this fix:
https://github.com/google/jimfs/issues/18

Test Plan:
Introduced a new unit test: `EdenProjectFilesystemDelegateTest`.
Incidentally, this is our first use of PowerMock in Buck.

Reviewed By: yiding

fbshipit-source-id: a472b95