Avoid deadlock caused by blocking xctool's stderr
Summary:
Read xctool's stderr in a separate thread to avoid a deadlock when
xctool tries to write more bytes on its stderr than fit in the
operating system's pipe buffer (tens of KiB on Mac OS X).

Test Plan:
Modify the test runner to output a lot of text on stderr, check
that Buck does not deadlock when running tests.

Reviewed By: Coneko

fbshipit-source-id: 368f080