Add support for Reason files.
Summary:
Reason files add two more file extensions: .`re` for implementations and `.rei` for interfaces.
The build rules need to recognize the new `.re` and `.rei` extensions.
Compilation commands need extra parameters in case of Reason files:
- `ocamlc` and `ocamlopt` need the Reason formatter `refmt` passed via the `-pp` option, and need to know whether files are implementation `-impl` or interface `-intf` files.
- `ocamldep` needs, for each file, to know whether it is an implementation or interface file, and whether an external processor needs to be used. Since once a processor is set, then it applies to all the subsequent files, and since Ocaml and Reason files can alternate, the `refmt` preprocessor needs to be sometimes set, and sometimes the default no-op formatter needs to be reset (currently by passing `cat` as the preprocessor).
- There are minor complication in the treatment of dependencies: `ocamldep` produces a dependency graph in terms of `.cmx` and `.cmi` files, and these need to be mapped back to source files. While with a project consisting only of Ocaml files the mapping is unique, with mixed Ocaml/Reason files there are two possible mappings, e.g. a `.cmx` file can map to a `.ml` or a `.re` file. The diff duplicates the dependency graph so that all the possibilities are covered, then uses filtering w.r.t. the input files to pick the correct ones.

The intent of the diff is to change as little as possible in Ocaml-only projects. This is partially possible, except for the dependency graph, which needs to take into account the possible existence of Reason files.

Test Plan:
Test on CI.
Tested separately on toy projects that mix `.ml` and `.re` files. Currently, I can't add them as extra tests until Reason is available on CI machines.

Reviewed By: Coneko

fbshipit-source-id: fb64e74