Move many ndk path calculations to a helper object
Summary:
Dealing with the fact that these paths are different for all
kinds of different ndk/compiler/runtime/architecture/etc is a huge pain,
particularly when updating the ndk. This is the first step to making it
much simpler.

There should be no behavior change and the paths are all calculated
exactly as before.

Test Plan: Run tests.

Reviewed By: dreiss, marcinkosiba

fbshipit-source-id: b67d86f