Add a script for generating artificial Buck repositories
Summary:
This adds a script that analyzes an existing Buck repository and
creates an artificial repository with specified number of targets
and similar properties (target type distribution, dependencies,
etc.) as the original repository. The script is currently able to
analyze a complex Android repository and create a repository that
can be parsed with `buck targets`.

Current drawbacks:
  * Not all target types are supported, the script will likely
    only work on Android repositories for now. This should be
    fairly simple to extend later.
  * Not all attributes of targets are supported, only the
    required minimum (all non-optional attributes).
  * All generated targets are currently places in the root of the
    repository in a single BUCK file.
  * The generated targets do not build since all generated sources
    are empty.
  * The script is very slow, mostly due to the way it generates
    files (trial and error directly on the file system).

Test Plan:
Run the script to generate a fake Android repository and confirm
that `buck targets` works in it.

Reviewed By: Coneko

fbshipit-source-id: 78ba549