Make it easier to configure different ndk paths
Summary:
Every ndk update, dealing with figuring out all
the paths calculated in NdkCxxPlatforms that are now wrong and fixing
them has been a huge pain (the process was like: do a build/test, see a
failure, look into compiler flags and try to figure out which really
long path was wrong, find that in NdkCxxPlatforms, add ugly/hard to
follow conditional path resolution). This change moves most of the
paths to be calculated off of a pattern like:
    "lib/{compiler_type}/{toolchain_target}/{compiler_version}"
And we assert that the resulting path actually exists. This makes it
much easier to figure out which path calculations are wrong and to
easily understand what paths are used with different ndk/compiler/etc
combinations and easily add a new such.

The paths are all calculated the same as before.

Test Plan: run tests

Reviewed By: marcinkosiba

fbshipit-source-id: 1e3db6f