use cached rulekey for build events instead of recomputing
Summary:
It seems like this buck event log code was recomputing the rule key
for every rule that passes through it instead of leveraging the cache
in RuleKeyBuilderFactory and co.

Fixing this seems to improve clean-build speed by a little. It should
be noted that the results should be profiled with yourkit probes turned
off, otherwise, they show a dramatic improvement.

Test Plan:
CI

Run a build such that everything is build, rerun the build to get the clean build time.

Reviewed By: k21

fbshipit-source-id: a738667