Allow nonexistant cell roots
Summary:
Cross-platform Buck projects can vary wildly to the point of one platform
relying on the targets of a cellroot that another platform doesn't.  Instead of
throwing errors when parsing BUCK files on platforms where this cell root doesn't
exist, let it succeed -- if the build really needs the cell root, it'll fail
when accessing targets in that root.

Test Plan:
Successfully run 'buck project' on MacOSX for a project with a Linux-only
cell root dependency.

Reviewed By: Coneko, mzlee

fbshipit-source-id: 9171261