actool, pass --target-device tv if tvOS
Summary:
We (Connected TV team) are working on a videos Apple TV app and I'm trying to get a build uploaded to TestFlight for initial team dogfooding. The App Store is happy with everything *except* it's complaining that the ipa doesn't include some necessary assets:

`ERROR ITMS-90471: "Missing Image Asset. Your app is missing the Home Screen Icon asset 'App Icon - Small' in 'Payload/FBTV.app'."`

I found the Assets.car file was missing the images and tracked it down to the actool invocation. As-is, it reports:

`App Icon - Small.imagestack: error: The image stack "App Icon - Small" must have at least 2 layers with applicable content. Although it has 3 layers, none have applicable content.`

but when I run with `--target-device tv` it's happy (and the resulting .car has the expected large, small and top shelf images as expected). Since the `applePlatformName` is unique proposing to use that so we only add this for tvOS

Test Plan:
* sandcastle
* `env NO_BUCKD=1 ~/buck/bin/buck build "fbtv#appletvos-arm64"` succeeds and the .car has the expected assets
* `env NO_BUCKD=1 ~/buck/bin/buck build fbeats` succeeds and has a Assets.car that looks correct

Reviewed By: yiding

fbshipit-source-id: 9682a2e