Generate artificial targets in multiple BUCK files
Summary:
Add a generator for the `visibility` field which will always set it
to `['PUBLIC']` so there are not restrictions on which targets can
depend on what. Performance-wise this should not be significantly
different from replicating the source repo's visibility usage.

Also analyze the distribution of number of targets in BUCK files
and create multiple BUCK files with the same distribution of
numbers of targets in them.

Test Plan: Run the script and check its output with `buck targets`.

Reviewed By: beefon

fbshipit-source-id: 93749c2