Generate artificial targets in a subdirectory
Summary:
With this change all targets are still generated in the same BUCK
file, but the BUCK file is no longer at the root of the generated
repository. This is the first step towards generating repositories
with multiple BUCK files.

Test Plan: Run the script, check output with `buck targets`.

Reviewed By: beefon

fbshipit-source-id: 3be94a2