C/C++: don't throw errors if duplicate shared libs are the same
Summary:
Previously, if the same library was pulled in via two different paths,
we'd throw an error complaining about this conflict.  This diff relaxes
the error a bit so that we only error out if the duplicate libraries
are different (as we do for headers) and would result in undefined
behavior.

Test Plan: added tests

Reviewed By: Coneko

fbshipit-source-id: aabcd18