Use case insensitive checks when generating files
Summary:
Always compare lowercase file names when checking whether a file
already exists in memory to guarantee that the generated files
can be created on a case-insensitive file system. With this we
can also remove the `try ... except` around the file creation.

Test Plan: Run the script, check the output with `buck targets`.

Reviewed By: Coneko

fbshipit-source-id: 7b3ff4d