Simplify dex creation for BuildConfig
Summary:
Previously, addBuildConfigDeps needed to know if we were pre-dexing, and
it generated the pre-dex rules itself.  Now, it just adds the
java_libraries that it creates to an ImmutableList.Builder passed in
from above.  The libraries are then treated as more-or-less normal
java_libraries and go through the normal code path to get their pre-dex
rules created.

This is less code, and will get even smaller with Java 8, but the real
reason I'm doing it is that it will make it easier to add a new
code-gen'ed java_library.

Test Plan:
Existing unit and integration tests cover this adequately, especially
MultipleBuildConfigIntegrationTest, which I ran manually to double-check
that the rules being generated had reasonable target names.

Reviewed By: marcinkosiba

fbshipit-source-id: 54115ea