Replace getTransitiveClasspathEntries with getTransitiveClasspaths
Summary:
Based on the observation that we only ever use the keys or values of
getTransitiveClasspathEntries, and never the map-like nature of the
entries themselves, this diff refactors all uses of the keySet to
`getTransitiveClasspathDeps` and all uses of value to the new
`getTransitiveClasspaths`.

This should dramatically reduce the memory footprint for storing all
the transitive classpaths.

The underlying computation to generate the classpaths still constructs
a multimap (because some codepaths actually use the underlying
`getClasspathEntries` function). I will contemplate removing this in
the future for a computation speed win.

Test Plan:
CI

`buck audit classpaths` shows no differences.

Reviewed By: alsutton

fbshipit-source-id: 99aea6e