Stop computing Multimap of transitive classpaths
Summary:
This diff adds a function to collect classpaths from libraries without
preserving the library it comes from (fetching the classpaths of a
particular library can be achieved by calling
getImmediateClasspaths()).

Change remaining uses of the multimap to use this new function, or
call getImmediateClasspaths.

This results in a negligible performance win when computing transitive
classpaths (~4% in one test).

Test Plan: CI, `buck audit classpaths` and see results are identical.

Reviewed By: Coneko

fbshipit-source-id: 059e7d4