export_file: introduce a now `mode` parameter to avoid copying
Summary:
Currently, the implementation of `export_file` creates a copy of
the references source path.  In many cases, this isn't necessary
and can incur a large build time overhead.  This diff adds a new
`mode` parameter which can be used to select a new `REFERENCE`
mode.  In this mode, the underlying source path is forwarded rather
than copied/cached.

Test Plan: added tests

Reviewed By: Coneko

fbshipit-source-id: 063f591