Ignore com.facebook.buck.intellij files from compilation by the main buck module.
Summary:
Don't compile the intellij module as part of the buck module.  It was
wrongly rooted, and even if we fix that, it has different dependencies
and should be treated differently -- as one example, it uses the
IntelliJ-compatible JDK, which is not necessarily the JDK used by
Buck.  (Buck is currently Java 7+, while the ideabuck plugin can,
independently of Buck, decide to run on Java 8+).

Test Plan: Verify the plugin files are no longer compiled as part of the buck module.

Reviewed By: asp2insp

fbshipit-source-id: 91e2c44