Refactor some logic out of MorePaths into MoreProjectFilesystems.
Summary:
Following the model of Guava, when we have a type XXX from a vendor (whose API
we do not control) and we have common utility methods we want to add for that
type, we introduce a utility class `MoreXXXs` for these methods. Most often,
the methods in the utility class take a parameter of type `XXX` as their first
argument.

In our `MorePaths` class, that idea has gotten away from us to the point that
one cannot use `MorePaths` without also taking a dependency on
`ProjectFilesystem`. This does not seem appropriate, so this moves those method
into their own, new class: `MoreProjectFilesystems`.

Alternatively, because we control the API of `ProjectFilesystem`, we should
just take the methods in `MoreProjectFilesystems` and add them to
`ProjectFilesystem` directly.

Test Plan: CI: Used IntelliJ to move the methods and then ran `buck autodeps` to clean everything up.

Reviewed By: Coneko

fbshipit-source-id: 84447f3