Make sure managed thread count respects --num-threads
Summary:
managed_thread_count originally was expected to be num_threads + number_of_cpu_cores, but formula was hardcoded incorrectly - it was just number_of_cpu_cores * 2.
Now we make it to be "relative" to the value of num_threads, so it will be computed like `num_threads + number_of_cpu_cores`.

Test Plan: CI

Reviewed By: Coneko

fbshipit-source-id: 5c880fd