handle the IOException for No space left on device.
Summary:
This is a very hacky way to make an IOException more readable to the user. The idea
here is that for specific IOException that we know why they happen but we also want
to keep them as IOException we can catch them at the top level, look at the message
and treat them based on that.

I can't say that this diff makes me happy, but it will do for now.

Test Plan:
manual testing when I made WriteFileStep.execute throw an exception with the correct
message.

Reviewed By: Coneko

fbshipit-source-id: 2a0aae3