Artificial project: not all java libraries have output
Summary:
To pass target to action graph transformation, we need to make sure
we only use targets with outputs as sources of other targets. This
works automatically in most cases, except when a single target type
has output in some cases and doesn't have output in other. An
example of this is java_library which only has output if it has at
least one source. To make sure we only use targets with outputs,
track those separately and use them when we need a source path.

Test Plan:
Run the script, check that `buck targets --show-rulekey` does not
complain about java_library without an input anymore.

Reviewed By: Coneko

fbshipit-source-id: ee5c23f