Decouple swift from apple platform
Summary:
Code base for supporting swift is getting larger, we should think about how to make buck to build swift platform-independently. This PR is a starting point for that, it is extracted from https://github.com/facebook/buck/pull/893, and should make it easier to specify swift version to be used with buck.
Closes https://github.com/facebook/buck/pull/896

Reviewed By: Coneko

fbshipit-source-id: 97c8b17