Reduce memory usage in HashingDeterministicJarWriter
Summary:
I noticed Buck was failing to build Buck occasionally, with
an `OutOfMemoryError`:

```
 java.lang.OutOfMemoryError: Java heap space
	at java.util.Arrays.copyOf(Arrays.java:3236)
	at java.io.ByteArrayOutputStream.toByteArray(ByteArrayOutputStream.java:191)
	at com.google.common.io.ByteStreams.toByteArray(ByteStreams.java:169)
	at com.facebook.buck.jvm.java.abi.StubJar$CreateStubAction.visit(StubJar.java:76)
	at com.facebook.buck.jvm.java.abi.ZipWalker.walk(ZipWalker.java:65)
	at com.facebook.buck.jvm.java.abi.StubJar.writeTo(StubJar.java:58)
	at com.facebook.buck.jvm.java.CalculateAbiStep.getPathToHash(CalculateAbiStep.java:73)
	at com.facebook.buck.jvm.java.CalculateAbiStep.execute(CalculateAbiStep.java:54)
```

I took a look, and found `StubJar` delegates its work to `HashingDeterministicJarWriter`.

Previously, `HashingDeterminsticJarWriter` took in a
`ByteSource` to provide the bytes to write to the JAR (and calculate
the hash code of).

However, since the `Walker` interface works in terms of `InputStream`,
we were previously reading the entire contents of the `InputStream`
into memory using `ByteSource.wrap(ByteStreams.toByteArray(stream)))`.

This is a lot of memory overhead, especially if the item we're writing to
the JAR is large.

A `ByteSource` is only useful if you intend to read the same file
multiple times. In this case, we want to:

1) Read the bytes of the file, and
2) (sometimes) Calculate a hash code

Thankfully, Guava provides a `HashingInputStream` class to do just this --
it wraps an `InputStream` and calculates a running hash as bytes are read
from the stream.

This means we don't need to use `ByteSource` as an API and can reduce the
number of copies.

Note that `HashingInputStream` closes the underlying stream when it's closed.
To make it consistent, I updated the other APIs in `HashingDeterministicJarWriter`
to close the `InputStream` passed in.

Test Plan: CI

Reviewed By: asp2insp

fbshipit-source-id: efebb05